+++ Source/Protocols/DBProfileViewControllerDataSource.h
@@ -27,6 +27,7 @@ NS_ASSUME_NONNULL_BEGIN
  @abstract Asks the data source to return the number of content controllers in the profile view controller.
  @param profileViewController The profile view controller requesting the information.
  @return The number of content controllers in the profile view controller.
+ @warning The number of content controllers must be greater than 0.
  */
 - (NSUInteger)numberOfContentControllersForProfileViewController:(DBProfileViewController *)profileViewController;
 
@@ -35,6 +36,7 @@ NS_ASSUME_NONNULL_BEGIN
  @param profileViewController The profile view controller requesting the information.
  @prarm index The index locating the content controller in the profile view controller.
  @return The content controller at a particular index in the profile view controller.
+ @warning The content controller cannot be nil.
  */
 - (DBProfileContentController *)profileViewController:(DBProfileViewController *)profileViewController contentControllerAtIndex:(NSUInteger)index;
 
@@ -43,8 +45,11 @@ NS_ASSUME_NONNULL_BEGIN
  @param profileViewController The profile view controller requesting the information.
  @prarm index The index locating the content controller in the profile view controller.
  @return The title for the content controller at a particular index in the profile view controller.
+ @warning The title cannot be nil.
  */
 - (NSString *)profileViewController:(DBProfileViewController *)profileViewController titleForContentControllerAtIndex:(NSUInteger)index;
+
+@optional
 
 /*!
  @abstract Asks the data source to return the subtitle for the content controller at a particular index in the profile view controller.