+++ Source/Views/DBProfileDetailsView.h
@@ -25,6 +25,7 @@ typedef NS_ENUM(NSInteger, DBProfileDetailsViewStyle)  {
 - (instancetype)initWithStyle:(DBProfileDetailsViewStyle)style;
 
 @property (nonatomic, strong, readonly) UIView *contentView;
+
 @property (nonatomic, assign) UIEdgeInsets contentInset;
 
 @property (nonatomic, strong, readonly) UILabel *nameLabel;